+++ changelog.md
@@ -9,6 +9,7 @@ See also the commit log at github: https://github.com/Arduino-IRremote/Arduino-I
 - Added functions stopTimer(), restartTimer() and restartTimerWithTicksToAdd().
 - Added rawlen and initialGap to IRData.
 - Added ReceiveAndSendHobToHood example.
+- Removed default value USE_DEFAULT_FEEDBACK_LED_PIN for last parameter of IRsend::begin(bool aEnableLEDFeedback, uint_fast8_t aFeedbackLEDPin).
 
 # 4.2.1
 - Fix wrong type of tEnableLEDFeedback in IRSend.hpp and IRReceive.hpp.